# specs/001-pos-inventory-ledger/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Moiz Mobile POS API
  version: 1.0.0
  description: >
    POS, inventory and customer ledger API for Moiz Mobile & Corporation, Danwran Lodhran.
    Every response uses the envelope described in conventions.md. All money is decimal with
    2 places; the server always recomputes totals from line items and ignores client totals.
servers:
  - url: /api
...
security:
  - bearerAuth: []
paths:
...
  /products:
    get:
      tags: [Products]
      summary: Search and list products
      description: >
        Single `search` parameter matches name, brand, model, category or barcode (FR-003).
        Staff receive ProductStaffDto (no costPrice); Admin receive ProductAdminDto.
      parameters:
        - { name: search, in: query, schema: { type: string } }
        - { name: category, in: query, schema: { type: string } }
        - { name: lowStockOnly, in: query, schema: { type: boolean, default: false } }
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/PageSize' }
      responses:
        '200':
          description: Paged products
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          items:
                            type: array
                            items:
                              oneOf:
                                - $ref: '#/components/schemas/ProductStaffDto'
                                - $ref: '#/components/schemas/ProductAdminDto'
    post:
      tags: [Products]
      summary: Create a product variant (Admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProductUpsertRequest' }
      responses:
        '201': { description: Created }
        '403': { $ref: '#/components/responses/Error' }
  /products/{id}:
    parameters: [ { $ref: '#/components/parameters/Id' } ]
    get:
      tags: [Products]
      summary: Get one product
      responses:
        '200': { description: Product }
        '404': { $ref: '#/components/responses/Error' }
    put:
      tags: [Products]
      summary: Update a product (Admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProductUpsertRequest' }
      responses:
        '200': { description: Updated }
    delete:
      tags: [Products]
      summary: Deactivate a product (Admin only) — never hard-deleted
      responses:
        '204': { description: Deactivated }
  /products/{id}/image:
    post:
      tags: [Products]
      summary: Upload a product image (Admin only, max 2 MB)
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file: { type: string, format: binary }
      responses:
        '200': { description: Image stored, returns relative path }
  /products/{id}/stock-movements:
    get:
      tags: [Products]
      summary: Stock movement history for a product (FR-005)
      parameters:
        - { $ref: '#/components/parameters/Id' }
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/PageSize' }
      responses:
        '200': { description: Paged movements }
  /products/{id}/adjust-stock:
    post:
      tags: [Products]
      summary: Manual stock adjustment (Admin only), audited
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [newQuantity, note]
              properties:
                newQuantity: { type: integer, minimum: 0 }
                note: { type: string, maxLength: 255 }
      responses:
        '200': { description: Adjusted }

  # ----------------------------------------------------------- Suppliers
...
  /expenses:
    get:
      tags: [Expenses]
      summary: List expenses (Admin only)
      parameters:
        - { name: categoryId, in: query, schema: { type: integer } }
        - { $ref: '#/components/parameters/From' }
        - { $ref: '#/components/parameters/To' }
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/PageSize' }
      responses:
        '200': { description: Paged expenses }
    post:
      tags: [Expenses]
      summary: Record an expense (Admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [categoryId, amount, expenseDate]
              properties:
                categoryId: { type: integer }
                amount: { type: number, format: double, minimum: 0.01 }
                expenseDate: { type: string, format: date-time }
                note: { type: string, maxLength: 255 }
      responses:
        '201': { description: Created }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64 }
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
    From:
      name: from
      in: query
      schema: { type: string, format: date-time }
    To:
      name: to
      in: query
      schema: { type: string, format: date-time }
  responses:
    Error:
      description: Error envelope
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Envelope' }
...
  schemas:
    Envelope:
      type: object
      required: [success]
      properties:
        success: { type: boolean }
        data: { nullable: true }
        error:
          nullable: true
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            traceId: { type: string }
            details:
              type: array
              items:
                type: object
                properties:
                  field: { type: string }
                  message: { type: string }
...
    ProductStaffDto:
      type: object
      description: Staff-facing product. Has NO costPrice or profit property by design (FR-040).
      properties:
        id: { type: integer }
        name: { type: string }
        category: { type: string }
        brand: { type: string, nullable: true }
        model: { type: string, nullable: true }
        barcode: { type: string, nullable: true }
        imagePath: { type: string, nullable: true }
        salePrice: { type: number, format: double }
        quantityOnHand: { type: integer }
        isLowStock: { type: boolean }
    ProductAdminDto:
      allOf:
        - $ref: '#/components/schemas/ProductStaffDto'
        - type: object
          properties:
            costPrice: { type: number, format: double }
            wholesalePrice: { type: number, format: double }
            retailPrice: { type: number, format: double }
            minStockThreshold: { type: integer }
            supplierId: { type: integer, nullable: true }
    ProductUpsertRequest:
      type: object
      required: [name, category, costPrice, salePrice, minStockThreshold]
      properties:
        name: { type: string, maxLength: 150 }
        category: { type: string, maxLength: 80 }
        brand: { type: string, maxLength: 80 }
        model: { type: string, maxLength: 80 }
        barcode: { type: string, maxLength: 64 }
        costPrice: { type: number, format: double, minimum: 0 }
        wholesalePrice: { type: number, format: double, minimum: 0 }
        retailPrice: { type: number, format: double, minimum: 0 }
        salePrice: { type: number, format: double, minimum: 0 }
        quantityOnHand: { type: integer, minimum: 0 }
        minStockThreshold: { type: integer, minimum: 0 }
        supplierId: { type: integer, nullable: true }
...